+++ src/collective/excelexport/upgrades.py
@@ -1,2 +1,0 @@
-def upgrade_to_2(context):
-    context.runImportStepFromProfile('collective.excelexport:default', 'plone.app.registry')